% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_nnn.R
\name{cube_identity}
\alias{cube_identity}
\title{Solved State of an N x N x N Cube}
\usage{
cube_identity(n)
}
\arguments{
\item{n}{Integer, the side of the cube}
}
\value{
Integer vector \code{1:(6*n^2)}
}
\description{
Solved State of an N x N x N Cube
}
\examples{
length(cube_identity(3))
length(cube_identity(4))
}
\seealso{
\code{\link{cube_is_colour_solved}}
}
