% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_nnn.R
\name{cube_move_names}
\alias{cube_move_names}
\title{Move Names of an N x N x N Cube}
\usage{
cube_move_names(n)
}
\arguments{
\item{n}{Integer, the side of the cube (at least 2)}
}
\value{
Character vector of \eqn{6n} move names, in table order
}
\description{
Move Names of an N x N x N Cube
}
\examples{
cube_move_names(2)
cube_move_names(3)
cube_move_names(4)
}
\seealso{
\code{\link{cube_moves}}
}
