% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_orbits.R
\name{cube_pieces_home}
\alias{cube_pieces_home}
\title{Pieces Home on a Cube}
\usage{
cube_pieces_home(state)
}
\arguments{
\item{state}{Integer vector of \eqn{6n^2} stickers.}
}
\value{
A named integer vector with elements \code{home} and \code{total}.
}
\description{
The blunt summary of \code{\link{cube_progress}}: how many pieces are home
out of how many there are.
}
\examples{
cube_pieces_home(cube_identity(4))   # all of them
}
\seealso{
\code{\link{cube_progress}}
}
