% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_group.R
\name{group_identity}
\alias{group_identity}
\title{Solved State of a Group}
\usage{
group_identity(group)
}
\arguments{
\item{group}{A \code{\link{perm_group}}}
}
\value{
Integer vector of length \code{n}
}
\description{
The identity, \code{1:n}, for whatever \code{n} the group works on.
}
\examples{
group_identity(topspin_group(6, 3))
length(group_identity(cube_group(3)))
}
