% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_group.R
\name{group_moves}
\alias{group_moves}
\title{Move Alphabet of a Group}
\usage{
group_moves(group)
}
\arguments{
\item{group}{A \code{\link{perm_group}}}
}
\value{
Character vector of move names
}
\description{
Move Alphabet of a Group
}
\examples{
group_moves(topspin_group(10, 4))
group_moves(cube_group(3))
}
