% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_group.R
\name{is_perm_group}
\alias{is_perm_group}
\title{Test Whether an Object is a Permutation Group}
\usage{
is_perm_group(x)
}
\arguments{
\item{x}{Any object}
}
\value{
Logical
}
\description{
Test Whether an Object is a Permutation Group
}
\examples{
is_perm_group(topspin_group(10, 4))
is_perm_group(1:10)
}
