% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{state_store_size}
\alias{state_store_size}
\title{Get State Store Size}
\arguments{
\item{xp}{External pointer to StateStore}
}
\value{
Integer, number of stored states
}
\description{
Returns the number of states currently stored in the StateStore.
}
\examples{
store <- create_state_store(6L)
state_store_size(store)
}
