% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_add_from_df}
\alias{store_add_from_df}
\title{Add States to Store from Data Frame}
\usage{
store_add_from_df(store, df, cycle_val)
}
\arguments{
\item{store}{External pointer to StateStore}

\item{df}{Data frame with V1..Vn columns plus metadata}

\item{cycle_val}{Integer, cycle number to assign}
}
\value{
Number of rows added (invisible)
}
\description{
Converts a data.frame/data.table of states (as returned by
\code{analyze_top_combinations}) into the compact C++ store.
}
