% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_combos_for_state}
\alias{store_combos_for_state}
\title{Find Combo Numbers Containing a State in a Cycle}
\usage{
store_combos_for_state(store, state, target_cycle)
}
\arguments{
\item{store}{External pointer to StateStore}

\item{state}{Integer vector}

\item{target_cycle}{Integer}
}
\value{
Integer vector of combo_numbers
}
\description{
Find Combo Numbers Containing a State in a Cycle
}
