% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_find_best_match}
\alias{store_find_best_match}
\title{Find Best Match by Manhattan Distance}
\usage{
store_find_best_match(store, target, candidate_indices = integer(0))
}
\arguments{
\item{store}{External pointer to StateStore}

\item{target}{Integer vector, target state}

\item{candidate_indices}{Integer vector of 0-based indices to search
(empty = search all)}
}
\value{
Integer, 0-based index of best match
}
\description{
Find Best Match by Manhattan Distance
}
