% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_find_intersections}
\alias{store_find_intersections}
\title{Find Intersections Between Two Stores}
\usage{
store_find_intersections(store_a, store_b)
}
\arguments{
\item{store_a}{External pointer to StateStore}

\item{store_b}{External pointer to StateStore}
}
\value{
Character vector of common state keys
}
\description{
Returns state keys present in both stores. O(min(N,M)) via hash lookup.
}
