% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_get_state}
\alias{store_get_state}
\title{Get State from Store}
\usage{
store_get_state(store, idx)
}
\arguments{
\item{store}{External pointer to StateStore}

\item{idx}{Integer, 0-based index}
}
\value{
Integer vector of length perm_length
}
\description{
Retrieves a single permutation state by 0-based index.
}
