% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_group}
\alias{store_group}
\title{Group Attached to a State Store}
\usage{
store_group(store)
}
\arguments{
\item{store}{External pointer to StateStore}
}
\value{
The \code{\link{perm_group}} the store was created with, or \code{NULL}
}
\description{
Group Attached to a State Store
}
\examples{
store_group(create_state_store(54L, group = cube_group(3)))
}
