% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_lookup}
\alias{store_lookup}
\title{Lookup State Indices by State Vector}
\usage{
store_lookup(store, state)
}
\arguments{
\item{store}{External pointer to StateStore}

\item{state}{Integer vector}
}
\value{
Integer vector of 0-based indices
}
\description{
Lookup State Indices by State Vector
}
