% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{store_set_opd}
\alias{store_set_opd}
\title{Set OPD Combo Filter for a Cycle}
\usage{
store_set_opd(store, target_cycle, combos)
}
\arguments{
\item{store}{External pointer to StateStore}

\item{target_cycle}{Integer, cycle number to filter}

\item{combos}{Integer vector of allowed combo_numbers}
}
\description{
Restricts indices_for_cycle and filter_middle_indices to only return
states from the specified combo_numbers for the given cycle.
}
