% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_store.R
\name{create_state_store}
\alias{create_state_store}
\title{Create a New State Store}
\usage{
create_state_store(perm_length, init_capacity = 10000L)
}
\arguments{
\item{perm_length}{Integer, length of each permutation state}

\item{init_capacity}{Integer, initial capacity (default 10000)}
}
\value{
External pointer to StateStore (XPtr)
}
\description{
Creates a C++ StateStore object for compact, incremental storage of
permutation states with hash-indexed lookup.
}
\examples{
store <- create_state_store(6L)
state_store_size(store)
}
